add validate api (#25)
this PR adds the basic verification APIs to statement definition and autogen verification in `@statement` based on statement declaration. Currently verification is split into `verify` which contains mandatory verification and `typecheck` which asserts possible runtime error after type inference. Future, we may want to give an option as `typecheck(allow_unknown=True)` so we skip type check if the type is currently not being figured out (leaving as `Any` or `TypeVar`)