further simplify interpreter and use a more clarified name (#118)
This PR removes the fallback attached to each method table for
associated dialects but instead the fallback is implemented inside the
corresponding interpreter. This is because I don't see any use case that
different dialect will use different fallback, and in fact no dialect is
using this feature right now.