explicit Python ast.Call lowering trait (#139)
this PR removes the default behaviour of `@statement` which creates a
lowering transform from `ast.Call` to the statement call. This is
because `Statement` constructor is not necessarily compatible with
`ast.Call` syntax, e.g when `Statement` contains a `Region` or `Block`,
while most `Statement` is just a special `call`.

The change here requires one explicitly declares `FromPythonCall` trait
in the statement definition, which thus triggers the lowering transform
to match a compatible syntax in Python AST. In the future, I think we
can consider improving duplication in trait declaration by using
inheritance, e.g how `BinOp` in `py.binop` dialect is currently
implemented.

This change is important for the following support of `with xxx as f`
syntax where a given statement could have a similar roughly meaning of
`with` syntax via a trait e.g `FromPythonWithSingleContext` indicates
the statement has a compatible syntax with `with ... as f:` syntax where
the statement's only region is treated as the with body.