fix circular imports (#143)
this PR fixes circular imports by removing the re-export of
`rewrite.abc` and re-organize `attrs` and `types`, which is slightly
breaking. More specifically,

- `types` is moved into `attrs` as `kirin.ir.attrs.types`
- `py.data` dialect is removed, and `PyAttr` is now part of builtin
attributes because it is used within the IR module
- `py.base` is added to carry the minimum lowering transforms
- `julia.emit` now has a default `PyAttr` emit rule.
- `rewrite.abc` is no longer re-exported to avoid circular imports

the re-exports are still kinda messy, we will do a round of refactor of
that in future PRs.