refactor fcf unroll (#147)
this PR address a few issues in previous #130 

1. this should be called unroll instead of inline because it does not
inline an existing stdlib function.
2. #130 only supports the case when collection is a constant, this
resolves the case when constant propagation fails (e.g `fn` is not pure)
but fails when collection has a constant length but unknown (which is
why this rewrite is called unroll)

so now this rewrite relys on type parameter of `IList` which indicates
the size of the immutable list, thus `Map` etc. can be unrolled without
knowning how this list was constructed. (thanks to type inference on
generics)

```python
%3 = ilist.Map(%fn, %coll)

# get unrolled to

%v_1 = indexing.GetItem(%coll, 0)
%3 = func.call(%fn, %v_0)
%v_2 = indexing.GetItem(%coll, 1)
%4 = func.call(%fn, %v_1)
%v_3 = indexing.GetItem(%coll, 2)
%5 = func.call(%fn, %v_2)
```

then in the case `%coll = ilist.New(...)` the `GetItem` inlining pass
will replace getitem with the `SSAValue` referring to the element
directly.


cc: @kaihsin