revise interpreter initialize interface (#166)
this fix #165 I'm making this change to make bloqade's implementation
simpler.

This PR changes the interpreter back to `dataclass`s because:

## Avoid manually forward `__init__` arguments/hints
we want to avoid manually forwarding type hints when inheriting from
`BaseInterpreter`, this is solved by `dataclass`.

## Separate initialization from construction
we want to separate interpreter global state initialization (initialize)
from interpreter construction (__init__), and now

1. `__init__` for declaring extra interpreter options per interpreter,
most of the case you just need to specify a `kw_only` field in the
interpreter
2. `initialize` for declaring how to initialize the interpreter state
per interpreter.

## Sanity improvements

### Interpreter class variables check

the interpreter subclass now checks if the required class variables are
specified in the subclass if the class is not declared as an `ABC`.

### More clarity on API naming convention

`eval` now always means evaluate an IR expression, e.g `eval_stmt` means
evaluate an IR expression and return a corresponding value that is
EQUIVALENT to what the IR means, e.g `1 + x` is evaluated to `2` when
the frame has state `x = 1`.

`run` now always refers evaluate an IR expression given a local state,
e.g `run_method` and `run_callable` asks for the input arguments of
`Method` or a callable `ir.Statement`. In the same vein, we rename
`eval` -> `run` (non-breaking but has been marked depreated)

### `run` now uses a lock to check if it is called recursively

`run` is the entry point of an interpreter which does some extra
housekeeping work, for calls inside the kernel, one should use
`run_callable` (to run callable statements) or `run_method` (to run a
method if the callee is runtime). Now if `run` is called inside a
`MethodTable`, we will throw an error saying no instead of complaining
we don't have a valid interpreter state.

## Code reusibility improvements

`interp.Err` is now removed as part of the result type, instead we will
just use Python exception handling (`try ... catch ...`) because we
should try to avoid two ways of doing the same thing (previously via
`interp.Err` or via `raise`) to throw an error. Besides doing a
rust-like result type in Python turns out not being pleasant without any
syntax sugar and we end up handling the `Err` object constantly using
`if else`.

Now one just throws the `interp.InterpreterError` or wrap an existing
Python exception in `interp.WrapException` the interpreter will catch
such error in the end and record the stacktrace.

cc: @weinbe58 @kaihsin this will likely result in some breaking changes
in analysis, no need to upgrade it right now. But it should simplify
things a bit. also let me know what you think, I'm gonna merge this on
main first.