refactor analysis with SSAValue hints (#234)
## Analysis now returns frame instead of `dict`

Previously we record the results into a global field `analysis.results`
by merging the `frame.entries` into this dictionary. This works for
simple cases, but in the case of constant prop, the purity flag is per
frame (e.g when you go through one of the successor branch/region).

As a result, when a function call has no return value, or a scf.For does
not have yield (thus no results). We will not mark it pure.

More generally, when there are extra results per function, we won't be
able to return that results. So now we return the frame we use to run
the analysis within, which contains everything one defines to collect
for the analysis.

## Fixing DCE on SCF

with the above change, we are able to fix DCE on SCF and DCE now only
checks `Pure` or `MaybePure`.

## Replacing `Hint` with `SSAValue.hints`

A new field `hints` is added to `SSAValue`. This is because I realize
there could be different kind of hints from analysis to carry over
within the IR. This also simplifies the type system making it less
complicated.

---

well this turns out to be quite breaking... But it simplifies things a
lot.

This also fixes a few bugs related to purity analysis - the purity
should be per frame/statement instead of per SSAValue (well... I always
felt this was strange, now it is correct). Constant prop is no longer
convoluted with purity analysis (still slightly but a lot easier to
decouple in the future with joint analysis).

As a result loops on pure functions can fold correctly now.