error if the statement do not belong to the interpretation (#288)
while I think the major work should be done in a round of validation.
But It will be annoying to check that everytime for a developer. So I
think we should check this during interpretation for the sake of
ergonomics. For performance users, the built-in interpreter is not an
option anyways.

cc: @weinbe58