remove verification within scf cons (#298)
these should be separate verifications on number of blocks in a region
because it is possible that at construction a `scf.IfElse` or `scf.For`
contains multiple blocks but simplifies to a single one afterwards. We
only check this when `verify` is called explicitly. This allows more
flexibility for rewrites e.g inlining within `if-else`/`for` body.

This also mark their region uses `SSACFG` convention.