Registry for `PyClass` values.  (#312)
This registry is required for things like Parsing so that there is some
way of mapping the string representation of a `PyClass` object to a Type
binding. For example when constructing a Function signature from the
string: `(type1, type2, ...) -> type3` I need to know how to construct
the type binding from the string representation. This PR also allows
ones to customize how `PyClass` is printed to avoid naming collisions
between different classes, this can be done either by a new prefix or a
custom display name for the type.

A side effect is that going forward it would be good to construct the
Type binding using this method directly. For example

```python
from ._dialect import dialect

class Qubit:
   pass

QubitType = dialect.register_py_class(Qubit, ...)

```

---------

Co-authored-by: Xiu-zhe (Roger) Luo <rogerluo.rl18@gmail.com>