refactor interp to clean up APIs + improve stacktrace (#369)
this PR cleans up the double `new_frame` API in the interpreter
framework (closes #363). We also refactor the `run_ssacfg_region` and
deprecate `run_block` interface to make linter happier (it was not
consistent in `AbstractInterpreter` vs `BaseInterpreter`).

We still need some further thinking on what is a value of `region` and
`block` (determine by their terminating statement/block?) so that
codegen framework can share more codebase with interpreter.