new code generator framework (#401)
This implements the new `EmitABC` based on the new interpreter
framework. While it does its job - emitting some target format in either
IO or str or whatever else.

It does not address the usecase that lowers between two Kirin dialect
groups, I'm thinking support this use case in the lowering framework
instead of emit because technically this is about constructing the same
Kirin IR based on another Kirin IR so we can just reuse the lowering
tools for constructing the Kirin IR and provide a visitor of the IR.
Otherwise it is hard to address cases like "replacing a statement with 3
statements"