Testing tools to compare Methods against lists of Statements (#441)
Add the `kirin.test` module containing useful assert functions for
writing tests that includes
- `assert_structurally_same` for comparing `Method` against a list of
`Statements`
- `assert_statements_same` for comparing `Statement` to other
`Statement` objects
These methods compare the attributes of statements, as well as the
arguments of the statements recursively.
Comparing the arguments can be turned off setting `check_args=False`.
Also has a `debug` option that prints out the statements in the method
one by one until it fails, so the user can identify on which statement
the Method fails to match the list of Statements.