Fix bug that ignored defs in if when missing from else (#417)
@Roger-luo this fixes the bug in #416 when there's a definition in the
`if` path, but not in the `else` path.

However, the issue with the `elif` persists since the definition from
the nested if isn't propagated upwards properly.

In code:

This fails on main, but works on this branch

```python
@kernel
def main(n: int):
    x = 0

    if x == n:
        x = 1
    else:
        y = 2  # noqa: F841

    return x

main.print()
```

This fails both on main and this branch:
```python
@kernel
def main_elif(n: int):
    x = 0

    if x == n:
        x = 3
    elif x == n + 1:
        x = 4

    return x

main_elif.print()
```

(or, equivalently)
```python
@kernel
def main_elif(n: int):
    x = 0

    if x == n:
        x = 3
    else:
        if x == n + 1:
            x = 4

    return x

```

We somehow need to lower the bodies first in order to obtain the nested
definitions in them, but I'm not sure how exactly.