Rewrite rule from scf to cf (#404)
this PR implements the rewrite rule from SCF to CF dialect.

This rule must be applied in a fixed point fashion and works by walking
a ssacfg region looking for `scf.For` and `scf.IfElse`. when if finds
these statements it inserts their body blocks and uses `cf` to create
the appropriate cfg for those statements. This only implements it for
one level of nesting; hence, it should be applied in a fixed-point
fashion.

---------

Co-authored-by: Roger-luo <rogerluo.rl18@gmail.com>
Co-authored-by: Xiu-zhe (Roger) Luo <rluo@quera.com>