Fix bug that caused error in type inference when returning from if (#427)
@weinbe58 @Roger-luo I'm still not 100% certain what the issue was, but
this fixes #402 .

I think it's because using the `frame` instead of creating a new one
meant it still had the function as `node` rather than the actual
`IfElse` statement. So we do need to create a new one here. Is that
correct?

Please take a look and let me know whether that explanation is correct,
I'd like to take this opportunity to learn ;)